Completed 2 implementations
Created implementation 1 of Helicopter with Roll, Pitch and Yaw functionality

Felt it was too "janky" and difficult to control

Created implementation 2 of Helicopter that simulates normal movement and tilt of helicopter with stabilization.

Much smoother implementation and easier to control, but lacks rolling functionality.

Added some comments to implementations, planning to add more comments

Planning to add roll functionality to implementation 2 somehow but may run short of time